Connect the api to a h2 console

+++ src/main/java/com/in28minutes/springboot/myfirstwebapp/todo/Todo.java
@@ -11,18 +11,22 @@ import jakarta.validation.constraints.Size;
 public class Todo {
 	@Id
 	@GeneratedValue
-	private long id;
+	private Integer id;
+	
 	private String username;
+	
 	@Size(min=10, message = "Descripition field must have at least 10 caracters")
 	private String description;
+	
 	private LocalDate targetDate;
+	
 	private boolean done;
 	
 	public Todo() {
 		
 	}
 	
-	public Todo(long id, String username, String description, LocalDate targetDate, boolean done) {
+	public Todo(Integer id, String username, String description, LocalDate targetDate, boolean done) {
 		super();
 		this.id = id;
 		this.username = username;
@@ -31,11 +35,11 @@ public class Todo {
 		this.done = done;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
